requirements.txt
cffi==1.5.2
cryptography==1.2.3
docutils==0.12
ecdsa==0.13
enum34==1.1.2
Fabric==1.10.2
idna==2.0
inflect==0.2.5
ipaddress==1.0.16
lockfile==0.12.2
ndg-httpsclient==0.4.0
paramiko==1.16.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==0.15.1
python-daemon==2.1.1
python-dateutil==2.5.2
PyYAML==3.11
requests==2.9.1
six==1.10.0
slackclient==1.0.0
websocket-client==0.35.0
yampy==1.0
